streamline source data tree nodes (inc. #174)
