+++ components/treeView/TreeNodeComponents.module.css
@@ -1,0 +1,3 @@
+.textNodeLink:hover .textName {
+  font-weight: 500;
+}
